feat!: added a 3 way jumpers (#28)
So while building the firmware for version 2, I found that I could not
easily build the firmware deterministically as a write to the EPROM is
required the first time you flash the firmware. It would be nice to
have a hardware option to decide what side it is. We solder anything
into one of two possible configurations anyways, thus this does not
make the PCB less flexible.

I would argue that this makes it more flexible as you don't need this
pin to get a working keyboard. You can still do it the old way if you
want to, although you'd have to change at most 5 lines in the firmware.
It just allows the keyboard to see what side it is without writing to
the EPROM the first time. Making the firmware process fully
reproducible.

I also added the same 3 way jumper to the back of the TRRS connector.
This way you have a resistor on each side pulling either the clock
or data line up to high.

BREAKING CHANGE: moved the top mounting hole so any previous cases are
incompatible. Here it provides better support and it is less in the way.

BREAKING CHANGE: Changed the order of the pins on the TRRS connector.
This results in a loss of interoperability to communicate with older
versions. This needed to be done to make routing easier, and cleaner.

Since this release already has a breaking change, I do not see the
problem with adding more as this version is already completely
incompatible with the previous release.

Closes #26: "Make jumpers to allow for assigning something to the left
or right side".